Add integration tests for Api

+++ PaymentGateway.ApplicationServices/IProcessPayment.cs
@@ -6,7 +6,7 @@ namespace PaymentGateway.ApplicationServices
 {
     public interface IProcessPayment
     {
-        Task<PaymentId> CreatePayment(string cardNumber, int ccv, int expiryMonthDate, int expiryYearDate, decimal amount, Currency currency);
+        Task<PaymentId> CreatePayment(string cardNumber, int ccv, int expiryMonthDate, int expiryYearDate, double amount, Currency currency);
         Task<Payment> ProcessPayment(PaymentId paymentId);
     }
 }
