+++ PaymentGateway.Read.Models/PaymentView.cs
@@ -5,11 +5,12 @@ namespace PaymentGateway.Read.Models
     public class PaymentView
     {
         [JsonIgnore]
-        public string  ETag       { get; set; }
-        public string  CardNumber { get; set; }
-        public string  Id         { get; set; }
-        public decimal Amount     { get; set; }
-        public string  Currency   { get; set; }
-        public string  Status     { get; set; }
+        public string ETag { get; set; }
+
+        public string CardNumber { get; set; }
+        public string Id         { get; set; }
+        public double Amount     { get; set; }
+        public string Currency   { get; set; }
+        public string Status     { get; set; }
     }
 }
